Añadido more interface

+++ myzombiebird/core/src/com/parazathy/myzombiebird/gameobjects/Bird.java
@@ -12,7 +12,7 @@ public class Bird {
 
 	private float rotation;
 	private int width;
-	private int height;
+	private float height;
 	
 	private float originalY;
